Added Mobile Formated Footer
